Remove BCSID from metadata key filters, and update TestKeyValueBlockIterator to not use it as a block data prefix
TestKeyValueBlockIterator does not care about the values of the prefixes it uses, it
only needs two different types of block prefixes, along with unprefixed blocks,
to run its tests. The existing tests were refactored to allow specifying an
arbitrary prefix to be created for testing, and make it easier to validate that
the iterator returned the correct blocks for each prefix.