HDDS-6869 Use UGI when checking Upgrade priv
When checking for admin privileges in the OM Upgrade path, the check is
only run against the short name representation of the UGI. The check
should be done against variations of the user name. Ideally, the check
should be against UGI Objects but the admin list created also takes in a
string representation of the user name via config.