Fix a corner case associated with S3 MultiPartUpload.
If S3 MPU has an invalid upload id, preExecute() will not fill in file encryption info.
This is intentional as later in validateAndUpdateCache(), prepareKeyInfo() will abort due to
invalid upload id, therefore it doesn't matter that file encryption info is missing. So we just
need to move the validation of file encryption info field after prepareKeyInfo().

Added a test to cover this corner case.