HDDS-11014. [hsync] Block finalization should also merge last chunk to blockDataTable.
When piggybacking is enabled along with incremental chunk list, closing a block output stream does not result in a standalone "close" request.
Therefore, the flag endOfBlock should be attached in the WriteChunk/PutBlock to let DataNode know the block is closing, so that the last chunk
can be merged into blockDataTable.

Change-Id: Idf5ecf75e784cf0824dea8b1fc152d9107f791cf